Precompile PegJS grammar to avoid CSP violations at runtime (#411)
I tried using the current master in Cloud UI, but it failed on CSP violation. This was due to pegjs compiling a grammar to JavaScript at runtime.

To avoid this, I've applied a Babel plugin that transforms the grammar through pegjs at build time. I also had to upgrade pegjs to work with the plugin. This also has the advantage that pegjs doesn't need to be shipped to the browser.