Fix modal min-width issue in IE11 (#1174)
IE has trouble with min-widths and flex boxes if not used as as a pixel value. This PR makes the min-width use our max-form width (400px) as the min-width we'd accept on a modal. I picked that value because it's good for short notices, and our modals tend to hold input values of some kind. I updated our default modal to not use style tags for width so this would be a better default for IE11.