Make TSLint issues be warnings, not errors, when running `src-docs` (#1538)
Now that we've merged #1529, all TypeScript files will trigger lint
warnings if their formatting needs updating with Prettier (which can be
done by fixing with the linter). So far, so good.

However, the Webpack config for `src-docs` pulls in the project's
TypesScript and TSLint settings, meaning that Webpack won't compile any TS
files that have lint problems. We don't need checking at this stage to be
so strict, because the pre-commit hook and general lint task already check
everything.

So, in order to stop docs development from being annoying, I've customised
the TSLint settings when running `src-docs` so that lint problems are now
warnings, instead of errors. They'll still have to get fixed before
committing.