[Wiki] React Router 5.x does not provide React Context (#1778)
This corrects the previous edit suggesting that `react-router` 5.0
was completely backward compatible with 4.x.

It turns out, while trying to use this trick to extract the `context`
within the constructor that `this.context.router` is `undefined`.

I _assume_ that we just need to use a context consumer, but I
do not see its name or type documented anywhere.