Fix EuiIcon so webpack can build dynamic require contexts (#1998)
The work to remove TSLint in favour of ESLint has the side effect of "fixing" a lint failure in EuiIcon. It changed a plain string concatenation into a string template. Normally this would be fine, but this particular string concatenation is used to load icons dynamically, and Webpack can't build a dynamic require context from the compiled version of the template string.

Revert the change and add a lint directive to stop it recurring.