[EuiIcon]: Add appendIconComponentCache function (#3481)
* [EuiIcon]:

* Add appendIconComponentCache function to complement the recently-added clearIconComponentCache function.

Use case is to work around issues with the dynamic loading: if an icon is not in the cache, an attempt is made to fetch it dynamically, however some environments (such as using certain bundlers, like Parcel: https://github.com/parcel-bundler/parcel/issues/112) do not transpile this functionality correctly.

Being able to manually manipulate the icon cache allows a work around for this issue.

* Fix appendIconComponentCache type, lint issues; add tests; add documentation

* [EuiIcon]: Implement suggested change to new appendIconComponentCache function: One or more icon(s) are passed in as an object of iconKey (string): IconComponent

* [EuiIcon]:
* Define an empty no-op mock of the `appendIconComponentCache` function in icon.testenv.tsx
* Update the main CHANGELOG.md with a summary of the new `appendIconComponentCache` EuiIcon feature.

Co-authored-by: Chandler Prall <chandler.prall@gmail.com>