Revert grid background color (#4542)
This reverts the background color of the data grid content. Although the light gray works well for loading states, it isn't a good experience for fixed height grids with a low amount of items. We'll solve the loading issue in a better way in 7.13. I'm leaving the scrollbar mixin as is because I think it will be useful in other spots.