Bump dependencies (#4808)
* trim 0.0.1 -> 0.0.3

Note: I had to add a yarn dependency resolution to upgrade this
transitive dependency because.

* Get rid of duplicate prop-types dependencies

There were four different versions of the prop-types package in the
yarn lockfile, I simply removed the three oldest ones.

* react-popper 1.0.2 -> 1.3.11

* Update date_picker snapshot

The InnerReference component appears to have changed as a result of
the react-popper upgrade in 7906201.